Added an optional path arg for preference writing
Added an arg to the `writeSourcePreferences` function
to allow for writing to a different directory in the project.

Changed the tests related to writing the config file to be more atomic;
now they write the test file into the `tests/` folder, and clean up the
file after each test is run.